requirements.txt
streamlit
pytesseract
gtts
langdetect
pillow
